moved users to datatable
